api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API service for website parser"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
celery = "^5.3.6"
redis = "^5.0.1"
aiohttp = "^3.9.3"
requests = "^2.31.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3 